Render socials validation
