--- js/components/renderSocials.js
@@ -1,18 +1,0 @@
-
-function renderSocials(data) {
-    console.log('Generuojamas socials turinys');
-    console.log(data);
-
-
-    let HTML = '';
-
-    for (let i = 0; i < data.length; i++) {
-        const item = data[i];
-        HTML += `<a href="${item.link}" target="_blank" class="fa fa-${item.icon}" aria-hidden="true"></a>`;
-        console.log(item.link, item.icon);
-    }
-
-    console.log(HTML);
-}
-
-export { renderSocials }
