+++ js/data/socialsData.js
@@ -4,31 +4,31 @@ const socialsData = [
     {
         icon: 'facebook',
         link: '#'
-},
-{
-    icon: 'twiter',
-    link: '#'
-},
-{
-    icon: 'dribbble',
-    link: '#'
-},
-{
-    icon: 'tumblr',
-    link: '#'
-},
-{
-    icon: 'linkedin',
-    link: '#'
-},
-{
-    icon: 'instagram',
-    link: '#'
-},
-{
-    icon: 'rss',
-    link: '#'
-}
-];
+    },
+    {
+        icon: 'twiter',
+        link: '#'
+    },
+    {
+        icon: 'dribbble',
+        link: '#'
+    },
+    {
+        icon: 'tumblr',
+        link: '#'
+    },
+    {
+        icon: 'linkedin',
+        link: '#'
+    },
+    {
+        icon: 'instagram',
+        link: '#'
+    },
+    {
+        icon: 'rss',
+        link: '#'
+    }
+    ];
 
 export { socialsData }
