+++ js/main.js
@@ -1,6 +1,6 @@
 // console.log('JS is running...');
 
-import { renderSocials } from "./components/renderSocials.js";
+import { renderSocials } from "./components/socials/renderSocials.js";
 import { socialsData } from './data/socialsData.js';
 
 renderSocials(socialsData);
